beautify code looks

+++ Python-projekty i zadania/Lab-1/Zadanie-5/script.py
@@ -1,5 +1,5 @@
-a = int(input('Podaj bok a: '))
-b = int(input('Podaj bok b: '))
+a = float(input('Podaj bok a: '))
+b = float(input('Podaj bok b: '))
 
 pole_prostokota = a * b
 obwod_prostokota = 2 * a + 2 * b
